feat(chat): support supergroup/topic n category

+++ database/migrations/2023_06_05_1685901673_create_chatrooms_table.php
@@ -10,7 +10,7 @@ class CreateChatroomsTable extends Migration
     {
         Schema::create('chatrooms', function (Blueprint $table) {
 
-            $table->integer('id');
+            $table->bigInteger('id');
             $table->string('type');
             $table->string('title');
             $table->primary('id');
